split tests one for service, and one for recording actions
